Aggregation interface skeleton

+++ src/main/java/com/medallia/data/Segment.java
@@ -1,5 +1,5 @@
 package com.medallia.data;
 
 public class Segment {
-	int[][] rawData;
+	long[][] rawData;
 }
