+++ src/main/java/com/medallia/type/FieldMapper.java
@@ -3,18 +3,18 @@ package com.medallia.type;
 import java.util.function.Function;
 
 public interface FieldMapper<T> {
-	T fromStorage(int val);
-	int toStorage(T val);
+	T fromStorage(long val);
+	long toStorage(T val);
 
-	static <T> FieldMapper<T> make(Function<T,Integer> toStorage, Function<Integer, T> from) {
+	static <T> FieldMapper<T> make(Function<T,Long> toStorage, Function<Long, T> from) {
 		return new FieldMapper<T>() {
 			@Override
-			public T fromStorage(int val) {
+			public T fromStorage(long val) {
 				return from.apply(val);
 			}
 
 			@Override
-			public int toStorage(T val) {
+			public long toStorage(T val) {
 				return toStorage.apply(val);
 			}
 		};
